Updates code to use ClanLib 3.0
git-svn-id: https://subversion.assembla.com/svn/adaptive_shooter@104 4bcb948d-9056-44c7-a9fa-a90fd10d4afe

Former-commit-id: 721dc86b396564d175c1ff7b661f2de63d606ae2
Former-commit-id: 65c42440d2a2391a3fcf9a8cb309c96d4aecdf64

+++ src/AdaptiveShooter/Menu.cpp
@@ -46,7 +46,7 @@ void Menu::update()
 
 	if (!_canExecute)
 	{
-		float dt = GameManager::getInstance()->getDeltaTime();
+		int dt = GameManager::getInstance()->getDeltaTime();
 		_inputeTimer += dt;
 
 		if (_inputeTimer >= _inputDelay)
